% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/squire_utils.R
\name{fit_spline_rt_khartoum}
\alias{fit_spline_rt_khartoum}
\title{Fit squire model using rw splines for khartoum}
\usage{
fit_spline_rt_khartoum(
  data,
  country,
  population,
  reporting_fraction,
  n_mcmc = 10000,
  replicates = 100,
  rw_duration = 14,
  hosp_beds = 1e+10,
  icu_beds = 1e+10,
  log_likelihood = NULL,
  initial_pars = NULL
)
}
\arguments{
\item{data}{Data to fit to.  This must be constructed with
\code{particle_filter_data}}

\item{country}{Character for country beign simulated. WIll be used to
generate \code{population} and \code{contact_matrix_set} if
unprovided. Either \code{country} or \code{population} and
\code{contact_matrix_set} must be provided.}

\item{population}{Population vector (for each age group). Default = NULL,
which will cause population to be sourced from \code{country}}

\item{reporting_fraction}{Reporting fraction. Numeric for what proportion of
the total deaths the reported deaths represent. E.g. 0.5 results in
the model calibrating to twice the deaths provided by \code{data$deaths}}

\item{n_mcmc}{number of mcmc mcmc iterations to perform}

\item{replicates}{number of trajectories (replicates) to be returned that are being sampled from the posterior probability results produced by \code{run_mcmc_chain}
to select parameter set. For each parmater set sampled, run particle filter with \code{n_particles} and sample 1 trajectory}

\item{rw_duration}{Random Walk/Spline Duration. Default = 14 days}

\item{hosp_beds}{General Hospital Beds}

\item{icu_beds}{ICU Beds}

\item{log_likelihood}{function to calculate log likelihood, must take named parameter vector as input,
allow passing of implicit arguments corresponding to the main function arguments.
Returns a named list, with entries:
  - $log_likelihood, a single numeric
  - $sample_state, a numeric vector corresponding to the state of a single particle, chosen at random,
  at the final time point for which we have data.
  If NULL, calculated using the function calc_loglikelihood.}

\item{initial_pars}{List of starting mcmc pars. Default = NULL}
}
\value{
Model fit from [squire:::pmcmc]
}
\description{
Fit squire model using rw splines for khartoum
}
