% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_utils.R
\name{generate_draws}
\alias{generate_draws}
\title{Generate draws using parameters drawn from posterior}
\usage{
generate_draws(
  out,
  pars_list,
  parallel = FALSE,
  log_likelihood = NULL,
  draws = NULL,
  interventions = NULL
)
}
\arguments{
\item{out}{Output of [[squire::pmcmc]]}

\item{pars_list}{Output of [[generate_parameters]]}

\item{parallel}{Are we simulating in parallel. Default = FALSE}

\item{log_likelihood}{Log likelihood function. Default = NULL}

\item{draws}{How many draws are being used from pars_list. Default = NULL,
which will use all the pars.}

\item{interventions}{Are new interventions being used or default. Default = NULL}
}
\description{
Generate draws using parameters drawn from posterior
}
